move the configured primitives dataclasses to a separate python file
